refactor: extract mf.py from networks/_legacy_full (3/N)
