+++ stisim/networks/__init__.py
@@ -1,3 +1,4 @@
 """STIsim networks package — split across base, mf, fsw, msm, layered_networks, matchers."""
 from .base import NoPartnersFound, BasePars, NetworkPars, BaseNetwork
+from .mf import MFPars, MFNetwork
 from ._legacy_full import *  # noqa: F401, F403 (remaining classes during refactor)
